# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Sys API server.",
        "title": "Sys API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/dashboard/overview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aggregates metrics, recommended mock scenario, competency radar, recent interviews, and syllabus roadmap.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get candidate dashboard overview",
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": models.DashboardOverviewResponse}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"error\": \"unauthorized\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/quizzes/daily-challenge": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the active daily architecture challenge question without revealing correct answer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Get today's daily architecture challenge",
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": models.QuizQuestion}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/quizzes/daily-challenge/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verifies the candidate's selected option and returns the explanation and correctness.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Verify daily challenge answer",
                "parameters": [
                    {
                        "description": "Challenge verification payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MonuChaudhary14_Archon_internal_models.VerifyDailyChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": models.VerifyDailyChallengeResponse}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/quizzes/decks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves available practice topic decks with candidate mastery completion percentages.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "List system design quiz decks",
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": []models.QuizDeckItem}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"error\": \"unauthorized\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/quizzes/decks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all questions for a selected deck without revealing correct answers.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Get questions for a quiz deck",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Deck ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": []models.QuizQuestion}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/quizzes/decks/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Evaluates candidate deck answers, updates database mastery, and returns a detailed question review.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Submit deck quiz answers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Deck ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submission payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MonuChaudhary14_Archon_internal_models.SubmitDeckQuizRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": models.SubmitDeckQuizResponse}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"error\": \"unauthorized\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_MonuChaudhary14_Archon_internal_models.SubmitDeckQuizRequest": {
            "type": "object",
            "required": [
                "answers"
            ],
            "properties": {
                "answers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "time_spent_seconds": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MonuChaudhary14_Archon_internal_models.VerifyDailyChallengeRequest": {
            "type": "object",
            "required": [
                "question_id",
                "selected_option_id"
            ],
            "properties": {
                "question_id": {
                    "type": "string"
                },
                "selected_option_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}